+++ Gemfile
@@ -1,3 +1,5 @@
+source "https://rubygems.org"
+
 ruby '2.6.1'
 
 gem 'ruby-debug-ide'
